Move Cloneable to its own file.
